feat(shs-6001): improve code organization for cards
