+++ docroot/themes/humsci/humsci_basic/src/js/shared/index.js
@@ -8,14 +8,14 @@ import './media/video-with-caption';
 import './animation/page-scroll';
 import './timeline/expand-collapse-timeline';
 import './accordion/accordion-toggle-all';
-import './linked-cards/linked-cards';
+import './cards/linked-cards';
+import './cards/structured-card';
 import './photo-album/colorbox';
 import './editoria11y/editoria11y';
 import './vertical-tabs/vertical-tabs';
 import './addtocal/addtocal';
 import './prefered-reduced-motion/prefered-reduced-motion';
 import './main-content-fallback/main-content-fallback';
-import './structured-card/structured-card';
 
 // PoC for new megamenu.
 import './megamenu/index';
